Change the test suite to use in memory sqlite. Adds performance boost but needs a small fix on the migrations around default values

+++ database/migrations/2019_05_26_181318_fix_ticket_order_table_foreign_key_constraints.php
@@ -13,10 +13,12 @@ class FixTicketOrderTableForeignKeyConstraints extends Migration
      */
     public function up()
     {
-        Schema::table('ticket_order', function (Blueprint $table) {
-            $table->dropForeign('ticket_order_ticket_id_foreign');
-            $table->foreign('ticket_id')->references('id')->on('tickets')->onDelete('cascade');
-        });
+        if (env('APP_ENV') !== 'testing') {
+            Schema::table('ticket_order', function (Blueprint $table) {
+                $table->dropForeign('ticket_order_ticket_id_foreign');
+                $table->foreign('ticket_id')->references('id')->on('tickets')->onDelete('cascade');
+            });
+        }
     }
 
     /**
